Avoiding recursive behaviour at locale errors
If the locale folder cannot be found, or the default language "en" is not present, this method will start a recursive loop with no exit criteria.
This change updates the code to first try and load the selected locale. It that fails, it will try to load default locale but only once. If that also fails, it will raise an alert and no longer hang the application.