Handing generic error messages from API
If an error response is received from the API, it can either contain a translation key OR a message shown to the end user as-is.

Just one example could be during "rename" if there is a conflict, and the backend can enlighten the user about this even if no translation key is present. If new translation keys are added, they will have precedence over the API response message according to the implementation of this commit.